LPD-70609 Rename AssistantMessageBalloon component to AIAssistantMessageBalloon
