+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/content_editor/components/AIAssistantChat/components/AIAssistantMessageBalloon.tsx
